added auth, finished crud

+++ src/service/data-service/category.js
@@ -1,12 +1,13 @@
 'use strict';
 
 const Sequelize = require(`sequelize`);
-const {Aliase} = require(`../../const`);
+const {Aliase, NOT_DROP_CATEGORY_MESSAGE} = require(`../../const`);
 
 
 module.exports = class CategoryService {
   constructor(sequelize) {
     this._Category = sequelize.models.Category;
+    this._Article = sequelize.models.Article;
     this._ArticleCategories = sequelize.models.ArticleCategories;
   }
 
@@ -45,10 +46,25 @@ module.exports = class CategoryService {
 
 
   async drop(id) {
+
+    // ??? не работает  onDelete: `restrict`, поэтому так
+    const articles = await this._ArticleCategories.findAll({
+      where: {'CategoryId': id}
+    });
+
+    if (articles && articles.length) {
+      return NOT_DROP_CATEGORY_MESSAGE;
+    }
+
     const category = await this._Category.destroy({
       where: {id}
     });
 
+    return !!category;
+  }
+
+  async update(id, data) {
+    const category = await this._Category.update(data, {where: {id}});
     return !!category;
   }
 };
